# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: TechSprout School LMS API
  version: 1.0.0
  description: >
    Production-grade REST API for TechSprout School LMS (Foundation & Security Phase P1).
    All routes versioned under /api/v1.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:3001/api/v1
    description: Local Development API Server
  - url: https://api.techsprout.edu/api/v1
    description: Production API Server
paths:
  /health:
    get:
      summary: System Health Check
      description: Returns health status of the API service and its backing database.
      operationId: getHealth
      tags:
        - Health
      security: []
      responses:
        '200':
          description: System operational
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '503':
          description: System degraded or database down
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /health/ready:
    get:
      summary: Readiness Probe
      description: Lightweight readiness check probe for orchestrators and load balancers.
      operationId: getReadiness
      tags:
        - Health
      security: []
      responses:
        '200':
          description: Application ready to accept traffic
          content:
            application/json:
              schema:
                type: object
                required:
                  - ready
                  - timestamp
                properties:
                  ready:
                    type: boolean
                    example: true
                  timestamp:
                    type: string
                    format: date-time

  /auth/register:
    post:
      summary: Register a new user
      description: Validates input, hashes password, assigns student role, and creates identity.
      operationId: registerUser
      tags:
        - Identity & Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error or duplicate account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

  /auth/login:
    post:
      summary: User Login
      description: Verifies credentials, issues HttpOnly session cookie, and returns user profile.
      operationId: loginUser
      tags:
        - Identity & Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
                example: techsprout_session=abc123xyz; Path=/; HttpOnly; SameSite=Lax
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

  /auth/google:
    get:
      summary: Initiate Google OAuth 2.0 flow
      description: Returns Google authorization URL and state parameter for client-initiated OAuth flow.
      operationId: initiateGoogleOAuth
      tags:
        - Identity & Authentication
      security: []
      parameters:
        - name: redirectUri
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Authorization initiation payload
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - url
                  - state
                properties:
                  success:
                    type: boolean
                    example: true
                  url:
                    type: string
                    format: uri
                  state:
                    type: string

  /auth/google/callback:
    get:
      summary: Google OAuth 2.0 GET Callback
      description: Handles Google OAuth authorization code redirect and issues session cookie.
      operationId: handleGoogleCallbackGet
      tags:
        - Identity & Authentication
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Authenticated via Google OAuth
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid authorization code or state mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

    post:
      summary: Google OAuth 2.0 POST Callback
      description: Alternative direct exchange endpoint for single-page applications.
      operationId: handleGoogleCallbackPost
      tags:
        - Identity & Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
                - state
              properties:
                code:
                  type: string
                state:
                  type: string
      responses:
        '200':
          description: Authenticated via Google OAuth
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid authorization code or state mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

  /auth/logout:
    post:
      summary: User Logout
      description: Revokes current session and clears the session cookie.
      operationId: logoutUser
      tags:
        - Identity & Authentication
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessResponse'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

  /auth/me:
    get:
      summary: Get Current User Session
      description: Returns the currently authenticated user's profile and roles.
      operationId: getCurrentUser
      tags:
        - Identity & Authentication
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: Authenticated user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

  /auth/otp/send:
    post:
      summary: Send Phone OTP
      description: Generates a 6-digit OTP code, stores it with 180s TTL, and dispatches via SMS gateway.
      operationId: sendOtp
      tags:
        - OTP & Phone Verification
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpSendRequest'
      responses:
        '200':
          description: OTP dispatched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessResponse'
        '400':
          description: Invalid phone number
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '429':
          description: Rate limit exceeded (max 3 per 15 mins)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

  /auth/otp/verify:
    post:
      summary: Verify Phone OTP and establish session
      description: Atomically verifies 6-digit OTP code, creates or links student user, and returns session.
      operationId: verifyOtp
      tags:
        - OTP & Phone Verification
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpVerifyRequest'
      responses:
        '200':
          description: OTP verified and session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid, expired, or already used OTP code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

  /admin/users:
    get:
      summary: List all users (Admin only)
      description: Returns list of registered users with role and status details.
      operationId: listUsers
      tags:
        - Admin
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserDto'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Forbidden (Requires Admin role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

  /admin/roles/assign:
    post:
      summary: Assign Role to User (Admin only)
      description: Assigns or revokes privileged roles. Generates immutable audit log record.
      operationId: assignRole
      tags:
        - Admin
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - role
              properties:
                userId:
                  type: string
                  format: uuid
                role:
                  type: string
                  enum: [student, admin]
      responses:
        '200':
          description: Role updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessResponse'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '403':
          description: Forbidden (Requires Admin role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

  /admin/audit-logs:
    get:
      summary: Retrieve Audit Logs (Admin only)
      description: Returns append-only system audit logs.
      operationId: getAuditLogs
      tags:
        - Admin
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Audit log entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLogDto'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: techsprout_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    RegisterRequest:
      type: object
      required:
        - name
        - username
        - email
        - password
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: 'Rafiqul Islam'
        username:
          type: string
          minLength: 3
          maxLength: 50
          pattern: '^[a-zA-Z0-9_]+$'
          example: 'rafiqul_islam'
        email:
          type: string
          format: email
          maxLength: 255
          example: 'rafiqul@techsprout.edu'
        phone:
          type: string
          pattern: '^(?:\\+8801|01)[3-9][0-9]{8}$'
          example: '01712345678'
        password:
          type: string
          minLength: 8
          maxLength: 100
          example: 'StrongP@ssword123'

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: 'student@techsprout.edu'
        password:
          type: string
          example: 'StudentPassword123!'

    OtpSendRequest:
      type: object
      required:
        - phone
      properties:
        phone:
          type: string
          pattern: '^(?:\\+8801|01)[3-9][0-9]{8}$'
          example: '01712345678'

    OtpVerifyRequest:
      type: object
      required:
        - phone
        - otp
      properties:
        phone:
          type: string
          pattern: '^(?:\\+8801|01)[3-9][0-9]{8}$'
          example: '01712345678'
        otp:
          type: string
          minLength: 6
          maxLength: 6
          pattern: '^[0-9]{6}$'
          example: '482910'

    UserDto:
      type: object
      required:
        - id
        - name
        - username
        - email
        - role
        - isVerified
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        phone:
          type:
            - string
            - "null"
        role:
          type: string
          enum: [student, admin]
        isVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type:
            - string
            - "null"
          format: date-time

    UserResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/UserDto'

    AuthResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: 'Authentication successful'
        user:
          $ref: '#/components/schemas/UserDto'
        token:
          type: string

    ApiSuccessResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          type: object
        requestId:
          type: string
        timestamp:
          type: string

    ApiErrorResponse:
      type: object
      required:
        - success
        - message
        - errorCode
        - statusCode
        - timestamp
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: 'Invalid credentials provided'
        errorCode:
          type: string
          example: 'AUTH_INVALID_CREDENTIALS'
        statusCode:
          type: integer
          example: 401
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          example: 'req_8f1b2c3d4e5f'
        details:
          type: object

    HealthResponse:
      type: object
      required:
        - status
        - timestamp
        - uptime
        - environment
      properties:
        status:
          type: string
          enum: [ok, degraded]
        timestamp:
          type: string
        uptime:
          type: number
        environment:
          type: string
        services:
          type: object
          properties:
            database:
              type: string
              enum: [up, down]
            redis:
              type: string
              enum: [up, down]

    AuditLogDto:
      type: object
      required:
        - id
        - action
        - targetType
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        actorId:
          type:
            - string
            - "null"
          format: uuid
        action:
          type: string
        targetType:
          type: string
        targetId:
          type:
            - string
            - "null"
        ipAddress:
          type:
            - string
            - "null"
        userAgent:
          type:
            - string
            - "null"
        requestId:
          type:
            - string
            - "null"
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time